add Lesson 1.2
